% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{age_data_setup}
\alias{age_data_setup}
\title{Set up age data from age distribution data}
\usage{
age_data_setup(age_width_years = c())
}
\arguments{
\item{age_width_years}{Vector of age width data in years}
}
\description{
Function which uses the age distribution data (width of each age category in years) from an age_data.txt file 
             to set up all static age-related parameter values
}
\details{
Takes in a vector of age width data and outputs a list of vectors of age parameter values
}
