% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{assert_bounded}
\alias{assert_bounded}
\title{Assert that x is between bounds}
\usage{
assert_bounded(
  x,
  left = 0,
  right = 1,
  inclusive_left = TRUE,
  inclusive_right = TRUE,
  name = deparse(substitute(x))
)
}
\arguments{
\item{x}{Value to verify as being within bounds}

\item{left}{Lower bound}

\item{right}{Upper bound}

\item{inclusive_left}{True/false indicator if lower bound is inclusive}

\item{inclusive_right}{True/false indicator if upper bound is inclusive}

\item{name}{Name}
}
\description{
x is between bounds
}
\details{
x is between bounds (inclusive or exclusive)
}
