% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{assert_file_exists}
\alias{assert_file_exists}
\title{Check that file exists at chosen path}
\usage{
assert_file_exists(
  x,
  message = "file not found at path \%s",
  name = deparse(substitute(x))
)
}
\arguments{
\item{x}{File location}

\item{message}{Message to output if file does not exist}

\item{name}{Name}
}
\description{
Check that file exists at chosen path
}
\details{
Check that file exists at chosen path
}
