% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{assert_in}
\alias{assert_in}
\title{Assert that x is a subset of y}
\usage{
assert_in(
  x,
  y,
  message = "all \%s must be in \%s",
  name_x = deparse(substitute(x)),
  name_y = nice_format(y)
)
}
\arguments{
\item{x}{Value(s) to check if subset of y}

\item{y}{Set of values x must be a subset of}

\item{message}{Message to output if not all x are in y}

\item{name_x}{Name of x}

\item{name_y}{Name of y}
}
\description{
x is a subset of y
}
\details{
x is a subset of y (all x are in y)
}
