% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{assert_int}
\alias{assert_int}
\title{Assert Integer}
\usage{
assert_int(
  x,
  message = "\%s must be integer valued",
  name = deparse(substitute(x))
)
}
\arguments{
\item{x}{Value to check if integer}

\item{message}{Message to output if x not integer}

\item{name}{Name}
}
\description{
x is integer
}
\details{
x is integer
}
