% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{assert_length}
\alias{assert_length}
\title{Assert that x is of a specified length n}
\usage{
assert_length(
  x,
  n,
  message = "\%s must be of length \%s",
  name = deparse(substitute(x))
)
}
\arguments{
\item{x}{Vector to check length of}

\item{n}{Required length of x}

\item{message}{Message to output if x is not of length n}

\item{name}{Name}
}
\description{
x is of a specified length n
}
\details{
x is of a specified length n
}
