% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{assert_list}
\alias{assert_list}
\title{Assert list}
\usage{
assert_list(x, message = "\%s must be a list", name = deparse(substitute(x)))
}
\arguments{
\item{x}{Value to check if list}

\item{message}{Message to output if x not a list}

\item{name}{Name}
}
\description{
x is a list
}
\details{
x is a list
}
