% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{assert_logical}
\alias{assert_logical}
\title{Assert logical}
\usage{
assert_logical(
  x,
  message = "\%s must be logical",
  name = deparse(substitute(x))
)
}
\arguments{
\item{x}{Value to check if logical}

\item{message}{Message to output if x not logical}

\item{name}{Name}
}
\description{
x is logical
}
\details{
x is logical
}
