% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{assert_non_null}
\alias{assert_non_null}
\title{x is not null}
\usage{
assert_non_null(
  x,
  message = "\%s cannot be null",
  name = deparse(substitute(x))
)
}
\arguments{
\item{x}{Value to check if not equal to null}

\item{message}{Message to output if x is equal to null}

\item{name}{Name}
}
\description{
Check that value x is not null
}
\details{
Check that x is not null
}
