% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{assert_not_in}
\alias{assert_not_in}
\title{Assert that none of x are in y}
\usage{
assert_not_in(
  x,
  y,
  message = "none of \%s can be in \%s",
  name_x = deparse(substitute(x)),
  name_y = nice_format(y)
)
}
\arguments{
\item{x}{Value(s) to check are not in y}

\item{y}{Set of values x must not be in}

\item{message}{Message to output if some of x are in y}

\item{name_x}{Name of x}

\item{name_y}{Name of y}
}
\description{
None of x are in y
}
\details{
None of x are in y
}
