% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{assert_null}
\alias{assert_null}
\title{x is null}
\usage{
assert_null(x, message = "\%s must be null", name = deparse(substitute(x)))
}
\arguments{
\item{x}{Value to check if equal to null}

\item{message}{Message to output if x is not equal to null}

\item{name}{Name}
}
\description{
Check that value x is null
}
\details{
Check that x is null
}
