% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{assert_numeric}
\alias{assert_numeric}
\title{Assert Numeric}
\usage{
assert_numeric(
  x,
  message = "\%s must be numeric",
  name = deparse(substitute(x))
)
}
\arguments{
\item{x}{Value to check if numeric}

\item{message}{Message to output if x not numeric}

\item{name}{Name}
}
\description{
x is numeric
}
\details{
x is numeric
}
