% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{assert_pos}
\alias{assert_pos}
\title{x is positive (with or without zero allowed)}
\usage{
assert_pos(
  x,
  zero_allowed = TRUE,
  message1 = "\%s must be greater than or equal to zero",
  message2 = "\%s must be greater than zero",
  name = deparse(substitute(x))
)
}
\arguments{
\item{x}{Value to check if positive}

\item{zero_allowed}{Indicator (TRUE/FALSE) of whether zero is allowed}

\item{message1}{Message to output if x is negative (when zero allowed)}

\item{message2}{Message to output if x is not positive (when zero disallowed)}

\item{name}{Name}
}
\description{
Check that value x is positive
}
\details{
Check that x is positive (zero allowed or disallowed using zero_allowed parameter)
}
