% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{assert_single}
\alias{assert_single}
\title{x is atomic and single valued (has length 1)}
\usage{
assert_single(
  x,
  message = "\%s must be a single value",
  name = deparse(substitute(x))
)
}
\arguments{
\item{x}{Value to check if atomic and single valued (has length 1)}

\item{message}{Message to output if x not atomic and single valued}

\item{name}{Name}
}
\description{
Check that value x is atomic and single valued (has length 1)
}
\details{
Check that x is atomic and single valued (has length 1)
}
