% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{assert_single_bounded}
\alias{assert_single_bounded}
\title{Assert Single bounded}
\usage{
assert_single_bounded(
  x,
  left = 0,
  right = 1,
  inclusive_left = TRUE,
  inclusive_right = TRUE,
  name = deparse(substitute(x))
)
}
\arguments{
\item{x}{value to check if single value bounded between limits}

\item{left}{left bound}

\item{right}{right bound}

\item{inclusive_left}{True/False is left bound inclusive?}

\item{inclusive_right}{True/False is left bound inclusive?}

\item{name}{Name}
}
\description{
x is single value bounded between limits
}
\details{
x is single value bounded between limits
}
