% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{assert_single_int}
\alias{assert_single_int}
\title{Assert Single integer}
\usage{
assert_single_int(x, name = deparse(substitute(x)))
}
\arguments{
\item{x}{Value to check if single integer}

\item{name}{Name}
}
\description{
x is single integer
}
\details{
x is single integer
}
