% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{assert_single_logical}
\alias{assert_single_logical}
\title{Assert single logical}
\usage{
assert_single_logical(x, name = deparse(substitute(x)))
}
\arguments{
\item{x}{Value to check if single logical}

\item{name}{Name}
}
\description{
x is single logical
}
\details{
x is single logical
}
