% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{assert_single_numeric}
\alias{assert_single_numeric}
\title{Assert Single Numeric}
\usage{
assert_single_numeric(x, name = deparse(substitute(x)))
}
\arguments{
\item{x}{Value to check if single numeric}

\item{name}{Name}
}
\description{
x is single numeric
}
\details{
x is single numeric
}
