% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{assert_single_pos_int}
\alias{assert_single_pos_int}
\title{x is single positive integer (with or without zero allowed)}
\usage{
assert_single_pos_int(x, zero_allowed = TRUE, name = deparse(substitute(x)))
}
\arguments{
\item{x}{Value to check if single positive integer}

\item{zero_allowed}{Indicator (TRUE/FALSE) of whether zero is allowed}

\item{name}{Name}
}
\description{
Check that value x is single positive integer
}
\details{
Check that x is single positive integer (zero allowed or disallowed using zero_allowed parameter)
}
