% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{assert_single_string}
\alias{assert_single_string}
\title{Assert String}
\usage{
assert_single_string(x, name = deparse(substitute(x)))
}
\arguments{
\item{x}{Value to check if single character string}

\item{name}{Name}
}
\description{
x is single character string
}
\details{
x is single character string
}
