% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{assert_string}
\alias{assert_string}
\title{Assert String}
\usage{
assert_string(
  x,
  message = "\%s must be character string",
  name = deparse(substitute(x))
)
}
\arguments{
\item{x}{Value to check if character string}

\item{message}{Message to output if x not character string}

\item{name}{Name}
}
\description{
x is character string
}
\details{
x is character string
}
