% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{assert_vector}
\alias{assert_vector}
\title{Assert Vector}
\usage{
assert_vector(
  x,
  message = "\%s must be a non-recursive vector",
  name = deparse(substitute(x))
)
}
\arguments{
\item{x}{Value to check if vector and not recursive}

\item{message}{Message to output if x not a non-recursive vector}

\item{name}{Name}
}
\description{
x is a vector (and is not a list or another recursive type)
}
\details{
x is a vector (and is not a list or another recursive type)
}
