% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cluster_input_setup}
\alias{cluster_input_setup}
\title{Cluster input setup}
\usage{
cluster_input_setup(age_file = "inst/extdata/Constant/age_data.txt",
  benchmark_file = "inst/extdata/Constant/results_example/Benchmark_details.txt",
  set_npt = 1, benchmark = "EIR", age_start = 0, age_end = 65)
}
\arguments{
\item{age_file}{File containing age distribution data}

\item{benchmark_file}{File containing benchmark data}

\item{set_npt}{Data point to use (1-max)}

\item{benchmark}{Benchmark type to use in choosing clusters ("EIR", "slide_prev", "pcr_prev", or "clin_inc")}

\item{age_start}{Starting age to use when calculating prevalence or incidence over age range (not used with EIR)}

\item{age_end}{End age to use when calculating prevalence or incidence over age range (not used with EIR)}
}
\description{
Function for taking benchmark data output by mainpop() and selecting the desired benchmark
             (EIR, slide prevalence, PCR prevalence, clinical incidence) data to use for input in
             setting up clusters, plus creating list of intervention parameter values
}
\details{
Takes in detailed benchmark data as a list (currently from a file) and outputs lists of chosen 
         benchmark values and intervention parameter values (currently as files)
}
