% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{clusters_create}
\alias{clusters_create}
\title{Create clusters}
\usage{
clusters_create(benchmark_file = "test_benchmark_list.txt",
  int_file = "test_int_list.txt", benchmark_mean = 0.25,
  benchmark_stdev = 0.025, int_mean = 0.15, int_stdev = 0.05)
}
\arguments{
\item{benchmark_file}{File containing benchmark data}

\item{int_file}{File containing intervention data}

\item{benchmark_mean}{Mean of benchmark value distribution}

\item{benchmark_stdev}{Standard deviation of benchmark value distribution}

\item{int_mean}{Mean of intervention parameter value distribution}

\item{int_stdev}{Standard deviation of intervention parameter value distribution}
}
\description{
Function for creating cumulative probability distributions and generating clusters from them
}
\details{
Takes in previously generated benchmark and intervention data (currently in files) and outputs cluster data
         (currently as files)
}
