% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{dataset_create}
\alias{dataset_create}
\title{Create new data set}
\usage{
dataset_create(
  dataset_folder = "",
  EIR_values = c(1),
  param_file = "",
  age_file = "",
  het_file = "",
  nyears = 10
)
}
\arguments{
\item{dataset_folder}{Folder where new dataset to be located. Will be created if not already in existence}

\item{EIR_values}{Vector of target values of annual entomological inoculation rate under constant rainfall conditions}

\item{param_file}{File containing model parameters to use for new dataset}

\item{age_file}{File containing age distribution parameters to use for new dataset}

\item{het_file}{File containing biting heterogeneity parameters to use for new dataset}

\item{nyears}{Number of years to run model for to achieve equilibrium}
}
\description{
Function which sets up a folder containing new data in the correct formats which can be used for
             input into mainpop().
}
\details{
Takes in a folder name, designated parameter files and a vector of target values of annual entomological 
             inoculation rate (EIR) under constant rainfall conditions, creates a folder containing parameter, starting data
             and annual EIR data files.
}
