% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ento}
\alias{ento}
\title{Entomological-only calculations}
\usage{
ento(
  input_folder = "",
  output_folder = NA,
  int_v_varied = 0,
  int_values = c(0),
  n_mv_set = c(),
  start_interval = 31,
  time_values = c(0, 7),
  flag_dt_adjust = 1
)
}
\arguments{
\item{input_folder}{Folder containing parameter files}

\item{output_folder}{Folder to send output files (no files saved if set to NA)}

\item{int_v_varied}{Intervention parameter given variable value 
(0= None, 1=ATSB kill rate, 2=bednet coverage, 3=IRS coverage)}

\item{int_values}{Vector of values of varied intervention parameter (will be unused if int_v_varied=0)}

\item{n_mv_set}{Vector of mosquito density number values to use (must be increasing order)}

\item{start_interval}{Period from start date before intervention starts (used to equilibrate lagged FOI data)}

\item{time_values}{Vector of time checkpoints}

\item{flag_dt_adjust}{Integer indicating whether or not to adjust dt by mosquito density in rcpp_mainpop
(0 = No, 1 = Yes)}
}
\description{
Function which takes in trial parameters for one or more mosquito populations, sends to C++ for 
computation of changes over time using deterministic continuum model, then collects and organizes results (TBA)
}
\details{
Takes a list of parameters, returns a list of raw data (data also saved to files as backup).
}
