% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{findPosition}
\alias{findPosition}
\title{Find position}
\usage{
findPosition(value = 0, set = c(0, 1))
}
\arguments{
\item{value}{Value to locate}

\item{set}{Set of values in which to locate value}
}
\description{
Find position of value in set
}
\details{
Find best position of value in set by finding closest value; modification of FindInterval() function
}
