% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_folder_data}
\alias{get_folder_data}
\title{Plot pre-calculated data from dataset folder}
\usage{
get_folder_data(
  input_folder = "",
  xvalues = "N_M",
  yvalues = "M",
  plot_flag = TRUE
)
}
\arguments{
\item{input_folder}{Dataset folder}

\item{xvalues}{Data to display on x-axis ("N_M" for line number, "M" for mosquito density parameter)}

\item{yvalues}{Data to display on y-axis ("M", "EIR", "slide_prev", "pcr_prev", or "clin_inc")
EIR represents annual total, prevalences and incidence represent year-round averages}

\item{plot_flag}{Logical operator indicating whether or not to plot graph of read values}
}
\description{
Function for displaying pre-calculated annual data or mosquito density parameter values from dataset folder
}
\details{
Reads data from annual_data.txt in dataset folder and outputs chosen benchmark values as a graph if desired; 
         returns list of x and y values as a list
}
