% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{load_dataset_folder}
\alias{load_dataset_folder}
\title{Load data set from folder}
\usage{
load_dataset_folder(dataset_folder = "")
}
\arguments{
\item{dataset_folder}{Dataset folder location}
}
\description{
Function which takes the name of a dataset folder and sets the locations for the files
}
\details{
Takes the location of a dataset folder, sets the locations for the files containing parameter values, 
             age and heterogeneity data, starting data and (if applicable) annual data and outputs them as a list for
             input to mainpop(). The filenames can also be set manually if they are not in a single dataset folder,
             e.g. if they are to be loaded from online
}
