% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{load_inputs}
\alias{load_inputs}
\title{Load input data from list of files}
\usage{
load_inputs(input_files = "", n_mv_set = c())
}
\arguments{
\item{input_files}{List of files containing parameter values, age and heterogeneity data, starting data, 
and (if applicable) annual data}

\item{n_mv_set}{Vector of mosquito density number values to use (must be increasing order)}
}
\description{
Function which takes a list of files and creates list of input data
}
\details{
Takes a list of files (as file locations or URLS, created by load_dataset_folder or set manually)
             and creates list of input data, including starting model data and parameter values
}
