% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mainpop}
\alias{mainpop}
\title{Main population evaluation}
\usage{
mainpop(rain_file = "Rainfall parameters.txt",
  parameter_file = "Simulation parameters.txt",
  input_file = "Starting data.txt", n_mv0_values = 600,
  n_mv0_start = 25, n_mv0_end = 50, int_v_varied = 1,
  n_int_values = 5, int_v_min = 0.05, int_v_max = 0.25,
  date_start = 180, date_int = 211, time_interval = 7, n_pts = 13,
  file_summary = "Benchmark_summary.txt",
  file_benchmarks = "Benchmark_details.txt",
  file_endpoints = "endpoints.txt", file_EIRd = "EIR.txt",
  file_imm_start = "imm.txt")
}
\arguments{
\item{rain_file}{File containing rainfall parameter values}

\item{parameter_file}{File containing simulation parameter values}

\item{input_file}{File containing mosquito and human input data}

\item{n_mv0_values}{Number of lines of data in input file}

\item{n_mv0_start}{Number of first set of data to use in input file}

\item{n_mv0_end}{Number of last set of data to use in input file (must be less than n_lines)}

\item{int_v_varied}{Intervention parameter given variable value (0= None, 1=ATSB kill rate, 2=bednet coverage, 3=IRS coverage)}

\item{n_int_values}{Number of values of varied intervention parameter to use (will be reset to 1 if int_v_varied=0)}

\item{int_v_min}{Minimum intervention parameter value (sole value used if n_int_values=1; unused if int_v_varied=0)}

\item{int_v_max}{Maximum intervention parameter value (unused if n_int_values=1 or int_v_varied=0)}

\item{date_start}{Day of the year when simulation starts (should be set based on input_file)}

\item{date_int}{Day of the year when intervention trial starts (period from date_start to date_int used to equilibrate lagged FOI data)}

\item{time_interval}{Length of time (in days) between measurements of benchmarks}

\item{n_pts}{Number of data points to take (one taken at date_int, so end time = date_int + ((n_pts-1)*time_interval))}

\item{file_summary}{Benchmark values at data capture points (sum over ages/heterogeneity categories where applicable)}

\item{file_benchmarks}{Benchmark values at data capture points (sum over heterogeneity categories where applicable)}

\item{file_endpoints}{Detailed endpoint data (only used if int_v_varied=0)}

\item{file_EIRd}{Daily EIR data during intervention period for each run}

\item{file_imm_start}{Immunity data at start of intervention for each mv_num value}
}
\description{
Function which takes in trial parameters, sends to C++ for computation, then collects and organizes results
}
\details{
Takes a list of parameters, returns a list of raw data (data also saved to files as backup).
}
