% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{nice_format}
\alias{nice_format}
\title{Nice Format}
\usage{
nice_format(x)
}
\arguments{
\item{x}{Value to convert to string}
}
\description{
for single value, return value as string. For vector of values return string of comma-separated values 
enclosed in curly brackets
}
\details{
x is single value or vector of values
}
