% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{plot_cohort_data}
\alias{plot_cohort_data}
\title{Display results of cohort simulations}
\usage{
plot_cohort_data(
  cohort_data = list(),
  benchmark = "slide_prev",
  flag_output = 1
)
}
\arguments{
\item{cohort_data}{List of form output by cohort() function}

\item{benchmark}{Benchmark type to output ("slide_prev", "pcr_prev", or "clin_inc")}

\item{flag_output}{Integer indicating how to present data (1: all clusters, 
2: average over all clusters)}
}
\description{
Function which takes the output of the cohort() function and displays selected results
}
\details{
Takes in a data frame of the form output by the cohort() function, calculates selected 
benchmark values and plots them on a graph
}
