% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{plot_folder_data}
\alias{plot_folder_data}
\title{Plot pre-calculated data from dataset folder}
\usage{
plot_folder_data(input_folder = "", xvalues = "N_M", yvalues = "M")
}
\arguments{
\item{input_folder}{Dataset folder}

\item{xvalues}{Data to display on x-axis ("N_M" for line number, "M" for mosquito density parameter)}

\item{yvalues}{Data to display on y-axis ("M", "EIR", "slide_prev", "pcr_prev", or "clin_inc")}
}
\description{
Function for displaying pre-calculated annual data or mosquito density parameter values from dataset folder
}
\details{
Reads data from annual_data.txt in dataset folder and outputs chosen benchmark values as a graph; returns graph 
         x and y values as a list
}
