% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{plot_mainpop_data}
\alias{plot_mainpop_data}
\title{Plot graph of main population data}
\usage{
plot_mainpop_data(input_list = list())
}
\arguments{
\item{input_list}{List containing main population benchmark data selected using get_mainpop_data}
}
\description{
Plot graph of main population data selected using get_mainpop_data function
}
\details{
Plot graph of main population data selected using get_mainpop_data function
}
