% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{tabulate_cohort_data}
\alias{tabulate_cohort_data}
\title{Tabulate results of cohort simulations}
\usage{
tabulate_cohort_data(cohort_data = list(), file = "dummy.txt")
}
\arguments{
\item{cohort_data}{List of form output by cohort() function}

\item{file}{Location/name of file to send results}
}
\description{
Function which takes the output of the cohort() function and tabulates results
}
\details{
Takes the output of the cohort() function and tabulates results by cluster, patient and time point for 
         processing using statistical models. Each patient across all clusters will have a unique ID number
         (patient_ID = (patient's number in their cluster) + (patient's cluster number - 1)*(number of patients per cluster))
}
