% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorpower.R
\docType{package}
\name{vectorpower}
\alias{vectorpower}
\alias{vectorpower-package}
\title{Power Calculation for Cluster Randomised Vector Control Trials}
\description{
Allows the user to estimate the sample size required for a given
  intervention efficacny in order to detect a significant effect, using data
  simulated from a cluster randomised trial in a malaria transmission model.
}
