% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cluster_input_setup}
\alias{cluster_input_setup}
\title{Cluster input setup}
\usage{
cluster_input_setup(
  input_list = list(),
  benchmark = "EIR",
  set_n_pt = 1,
  set_n_int = 1,
  age_start = 0,
  age_end = 65,
  plot_flag = TRUE
)
}
\arguments{
\item{input_list}{List containing mainpop output data}

\item{benchmark}{Benchmark to use in choosing clusters ("EIR", "slide_prev", "pcr_prev", or "clin_inc"
for new data, "EIR_annual", "slide_prev_annual", "pcr_prev_annual", "clin_inc_annual" for
pre-existing annual data in input folder)}

\item{set_n_pt}{Data point to use (1-max)}

\item{set_n_int}{Intervention number to use (1-max)}

\item{age_start}{Starting age to use when calculating prevalence or incidence over age range 
(not used with EIR)}

\item{age_end}{End age to use when calculating prevalence or incidence over age range 
(not used with EIR)}

\item{plot_flag}{Logical operator indicating whether or not to plot graph of read values}
}
\description{
Function for taking benchmark data output by mainpop() and selecting the desired benchmark
             (EIR, slide prevalence, PCR prevalence, clinical incidence) data to use for input in
             setting up clusters, plus creating list of intervention parameter values
}
\details{
Takes in detailed benchmark data as a list and outputs lists of chosen 
         benchmark values and intervention parameter values (as both list and files)
}
