% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{clusters_create}
\alias{clusters_create}
\title{Create clusters}
\usage{
clusters_create(
  input_list = list(),
  n_clusters = 100,
  benchmark_mean = 0.25,
  benchmark_stdev = 0.025,
  int_mean = 0.15,
  int_stdev = 0.05,
  plot_flag = FALSE
)
}
\arguments{
\item{input_list}{List containing input_data, produced by cluster_input_setup()}

\item{n_clusters}{Number of clusters to create}

\item{benchmark_mean}{Mean of benchmark value distribution}

\item{benchmark_stdev}{Standard deviation of benchmark value distribution}

\item{int_mean}{Mean of intervention parameter value distribution}

\item{int_stdev}{Standard deviation of intervention parameter value distribution}

\item{plot_flag}{True/false flag indicating whether to plot graphs of cluster data}
}
\description{
Function for creating cumulative probability distributions and generating clusters from them
}
\details{
Takes in previously generated benchmark and intervention data (currently in files) and outputs cluster data
         (currently as files)
}
