% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cohort}
\alias{cohort}
\title{Cohort evaluation}
\usage{
cohort(
  mainpop_data = list(),
  cluster_data = list(),
  n_patients = 100,
  output_folder = NA,
  prop_T_c = 0.9,
  age_start = 0.5,
  age_end = 10
)
}
\arguments{
\item{mainpop_data}{List output by mainpop() containing main population data}

\item{cluster_data}{List output by clusters_create() containing cluster data}

\item{n_patients}{Number of trial cohort patients per cluster}

\item{output_folder}{Folder to send results data (set to NA to omit saving to file)}

\item{prop_T_c}{Treatment probability in cohort}

\item{age_start}{Minimum age of trial cohort patients}

\item{age_end}{Maximum age of trial cohort patients}
}
\description{
Run stochastic individual model across trial cohort individuals in one or more clusters
}
\details{
Takes a list of parameters, returns a list of raw data (data also saved to files as backup).
}
