% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_mainpop_data}
\alias{get_mainpop_data}
\title{Get selected main population output data as a function of time}
\usage{
get_mainpop_data(
  input_list = list(),
  benchmark = "EIR",
  set_n_mv = 1,
  set_n_int = 1,
  age_start = 0,
  age_end = 65
)
}
\arguments{
\item{input_list}{List containing mainpop output data}

\item{benchmark}{Benchmark type to use in choosing clusters ("EIR", "slide_prev", "pcr_prev", or "clin_inc")
Represents annual total in case of EIR, year-round average for others}

\item{set_n_mv}{Baseline malaria level number(s) to output data for (from numbers in main population data, not
original data set)}

\item{set_n_int}{Intervention number(s) to output data for (from numbers in main population data)}

\item{age_start}{Starting age to use when calculating prevalence or incidence over age range (not used with EIR)}

\item{age_end}{End age to use when calculating prevalence or incidence over age range (not used with EIR)}
}
\description{
Takes in detailed benchmark data as a list and  returns selected benchmark values as a list
}
\details{
Function for taking output of mainpop(), selecting the desired benchmark
         (EIR, slide prevalence, PCR prevalence, clinical incidence) data, and outputting
         values of the benchmark as a function of time for the desired baseline malaria level(s)
         and intervention parameter value(s)
}
