% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mainpop}
\alias{mainpop}
\title{Main population evaluation}
\usage{
mainpop(input_folder = "inst/extdata/Constant/",
  output_folder = "inst/extdata/Constant/results_example/",
  n_lines = 1, n_mv_start = 0, n_mv_end = 0, int_v_varied = 1,
  n_int_values = 1, int_v_min = 0, int_v_max = 0, date_start = 0,
  date_int = 31, time_interval = 7, n_pts = 2)
}
\arguments{
\item{input_folder}{Folder containing rainfall, model parameter and starting data}

\item{output_folder}{Folder to send output files}

\item{n_lines}{Number of lines of data in input file}

\item{n_mv_start}{Number of first set of data to use in input file}

\item{n_mv_end}{Number of last set of data to use in input file (must be less than n_lines)}

\item{int_v_varied}{Intervention parameter given variable value 
(0= None, 1=ATSB kill rate, 2=bednet coverage, 3=IRS coverage)}

\item{n_int_values}{Number of values of varied intervention parameter to use (will be reset to 1 if int_v_varied=0)}

\item{int_v_min}{Minimum intervention parameter value (sole value used if n_int_values=1; unused if int_v_varied=0)}

\item{int_v_max}{Maximum intervention parameter value (unused if n_int_values=1 or int_v_varied=0)}

\item{date_start}{Day of the year when simulation starts (should be set based on input_file)}

\item{date_int}{Day of the year when intervention trial starts 
(period from date_start to date_int used to equilibrate lagged FOI data)}

\item{time_interval}{Length of time (in days) between measurements of benchmarks}

\item{n_pts}{Number of data points to take 
(one taken at date_int, so end time = date_int + ((n_pts-1)*time_interval))}
}
\description{
Function which takes in trial parameters for one or more populations, sends to C++ for computation 
of changes over time using deterministic continuum model, then collects and organizes results (EIR over time + 
starting immunity data for intervention computation or endpoint population data for control population, 
benchmark values for both)
}
\details{
Takes a list of parameters, returns a list of raw data (data also saved to files as backup).
}
