% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mainpop}
\alias{mainpop}
\title{Main population evaluation}
\usage{
mainpop(
  input_folder = "inst/extdata/Constant/",
  output_folder = NA,
  n_mv_set = c(1),
  int_v_varied = 0,
  int_values = c(0),
  start_interval = 31,
  time_values = c(0, 7)
)
}
\arguments{
\item{input_folder}{Folder containing rainfall, model parameter and starting data}

\item{output_folder}{Folder to send output files (no files saved if set to NA)}

\item{n_mv_set}{Vector of mosquito density number values to use (must be increasing order)}

\item{int_v_varied}{Intervention parameter given variable value 
(0= None, 1=ATSB kill rate, 2=bednet coverage, 3=IRS coverage)}

\item{int_values}{Vector of values of varied intervention parameter (will be unused if int_v_varied=0)}

\item{start_interval}{Period from start date before intervention starts (used to equilibrate lagged FOI data)}

\item{time_values}{Vector of time checkpoints}
}
\description{
Function which takes in trial parameters for one or more populations, sends to C++ for computation 
of changes over time using deterministic continuum model, then collects and organizes results (EIR over time + 
starting immunity data for intervention computation or endpoint population data for control population, 
benchmark values for both)
}
\details{
Takes a list of parameters, returns a list of raw data (data also saved to files as backup).
}
