% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_mainpop_data}
\alias{plot_mainpop_data}
\title{Plot selected main population output data as a function of time}
\usage{
plot_mainpop_data(input_list = list(), benchmark = "EIR",
  set_n_int = 1, age_start = 0, age_end = 65)
}
\arguments{
\item{input_list}{List containing mainpop output data}

\item{benchmark}{Benchmark type to use in choosing clusters ("EIR", "slide_prev", "pcr_prev", or "clin_inc")}

\item{set_n_int}{Intervention number to use (1-max)}

\item{age_start}{Starting age to use when calculating prevalence or incidence over age range (not used with EIR)}

\item{age_end}{End age to use when calculating prevalence or incidence over age range (not used with EIR)}
}
\description{
Function for taking output of mainpop(), selecting the desired benchmark
             (EIR, slide prevalence, PCR prevalence, clinical incidence) data, and 
             plotting it on a graph as an aid to setting up clusters
}
\details{
Takes in detailed benchmark data as a list and plots a graph based on the
         input parameters
}
